Implementando as exceções

+++ src/main/java/com/yuricsg/crud/service/ClientService.java
@@ -3,6 +3,8 @@ package com.yuricsg.crud.service;
 import com.yuricsg.crud.dto.ClientDTO;
 import com.yuricsg.crud.entities.Client;
 import com.yuricsg.crud.repository.ClientRepository;
+import com.yuricsg.crud.service.exception.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -20,7 +22,8 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id){
-        Client client = repository.findById(id).get();
+        Client client = repository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Cliente não encontrado"));
         return new ClientDTO(client);
     }
 
@@ -45,20 +48,27 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto){
+        try{
+            Client entity = repository.getReferenceById(id);
+            entity.setName(dto.getName());
+            entity.setCpf(dto.getCpf());
+            entity.setIncome(dto.getIncome());
+            entity.setBirthDate(dto.getBirthDate());
+            entity.setChildren(dto.getChildren());
 
-        Client entity = repository.getReferenceById(id);
-        entity.setName(dto.getName());
-        entity.setCpf(dto.getCpf());
-        entity.setIncome(dto.getIncome());
-        entity.setBirthDate(dto.getBirthDate());
-        entity.setChildren(dto.getChildren());
-
-        entity = repository.save(entity);
-        return new ClientDTO(entity);
+            entity = repository.save(entity);
+            return new ClientDTO(entity);
+        }
+        catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Não foi possível atualizar");
+        }
     }
 
     @Transactional
     public void delete(Long id){
+       if(!repository.existsById(id)){
+           throw new ResourceNotFoundException("Não foi possível deletar.");
+       }
         repository.deleteById(id);
     }
 }
